报名

+++ src/main/java/com/eco/dao/BaseDao.java
@@ -11,7 +11,7 @@ public interface BaseDao<T> {
 	/*
 	 * 返回数据库自增主键
 	 * */
-	int insert(String sql, Object... params);
+	Long insert(String sql, Object... params);
 	
 	/*
 	 * 返回 insert, delete，update影响行数
